pyproject.toml
[project]
name = "technews-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apscheduler>=3.11.2",
    "discord-py>=2.7.1",
    "feedparser>=6.0.12",
    "google-genai>=1.47.0",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "selectolax>=0.4.7",
    "supabase>=2.28.3",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
]
